+++ R/print.r
@@ -8,7 +8,7 @@
 #' @examples
 #'     library(rbiom)
 #'     
-#'     infile <- system.file("extdata", "hmp50.biom", package = "rbiom")
+#'     infile <- system.file("extdata", "hmp50.bz2", package = "rbiom")
 #'     biom <- read.biom(infile)
 #'     
 #'     print(biom)