+++ R/rarefy.r
@@ -21,7 +21,7 @@
 #' @examples
 #'     library(rbiom)
 #'
-#'     infile <- system.file("extdata", "hmp50.biom", package = "rbiom")
+#'     infile <- system.file("extdata", "hmp50.bz2", package = "rbiom")
 #'     biom <- read.biom(infile)
 #'     range(slam::col_sums(biom$counts))
 #'