+++ R/select.r
@@ -23,7 +23,7 @@
 #' @examples
 #'     library(rbiom)
 #'     
-#'     infile <- system.file("extdata", "hmp50.biom", package = "rbiom")
+#'     infile <- system.file("extdata", "hmp50.bz2", package = "rbiom")
 #'     biom <- read.biom(infile)
 #'     
 #'     ex1 <- select(biom, c('HMP14', 'HMP22', 'HMP03'))